@@ -1,18 +1,47 @@
 module github.com/folbricht/routedns
 
-go 1.15
+go 1.18
 
 require (
-	github.com/BurntSushi/toml v0.3.1
-	github.com/heimdalr/dag v1.0.1
+	github.com/BurntSushi/toml v1.2.1
+	github.com/RackSec/srslog v0.0.0-20180709174129-a4725f04ec91
+	github.com/heimdalr/dag v1.2.1
 	github.com/jtacoma/uritemplates v1.0.0
-	github.com/lucas-clemente/quic-go v0.22.0
-	github.com/miekg/dns v1.1.43
-	github.com/oschwald/maxminddb-golang v1.8.0
-	github.com/pion/dtls/v2 v2.0.9
+	github.com/miekg/dns v1.1.50
+	github.com/oschwald/maxminddb-golang v1.10.0
+	github.com/pion/dtls/v2 v2.1.5
 	github.com/pkg/errors v0.9.1
-	github.com/sirupsen/logrus v1.8.1
-	github.com/spf13/cobra v1.1.3
-	github.com/stretchr/testify v1.7.0
-	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d
+	github.com/quic-go/quic-go v0.32.0
+	github.com/sirupsen/logrus v1.9.0
+	github.com/spf13/cobra v1.6.1
+	github.com/stretchr/testify v1.8.1
+	golang.org/x/net v0.4.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/emirpasic/gods v1.18.1 // indirect
+	github.com/go-task/slim-sprig v0.0.0-20210107165309-348f09dbbbc0 // indirect
+	github.com/golang/mock v1.6.0 // indirect
+	github.com/google/pprof v0.0.0-20221219190121-3cb0bae90811 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/onsi/ginkgo/v2 v2.6.1 // indirect
+	github.com/pion/logging v0.2.2 // indirect
+	github.com/pion/transport v0.14.1 // indirect
+	github.com/pion/udp v0.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/quic-go/qpack v0.4.0 // indirect
+	github.com/quic-go/qtls-go1-18 v0.2.0 // indirect
+	github.com/quic-go/qtls-go1-19 v0.2.0 // indirect
+	github.com/quic-go/qtls-go1-20 v0.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/crypto v0.4.0 // indirect
+	golang.org/x/exp v0.0.0-20221227203929-1b447090c38c // indirect
+	golang.org/x/mod v0.7.0 // indirect
+	golang.org/x/sys v0.3.0 // indirect
+	golang.org/x/text v0.5.0 // indirect
+	golang.org/x/tools v0.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
