@@ -2,14 +2,15 @@ package rdns
 
 import (
 	"crypto/tls"
-	"io/ioutil"
+	"encoding/binary"
+	"io"
 	"net"
 	"sync"
 	"time"
 
-	quic "github.com/lucas-clemente/quic-go"
 	"github.com/miekg/dns"
 	"github.com/pkg/errors"
+	quic "github.com/quic-go/quic-go"
 	"github.com/sirupsen/logrus"
 )
 
@@ -26,7 +27,7 @@ type DoQClient struct {
 	log      *logrus.Entry
 	metrics  *ListenerMetrics
 
-	session doqSession
+	connection doqConnection
 }
 
 // DoQClientOptions contains options used by the DNS-over-QUIC resolver.
@@ -78,7 +79,7 @@ func NewDoQClient(id, endpoint string, opt DoQClientOptions) (*DoQClient, error)
 		DoQClientOptions: opt,
 		requests:         make(chan *request),
 		log:              log,
-		session: doqSession{
+		connection: doqConnection{
 			hostname:  host,
 			endpoint:  endpoint,
 			lAddr:     lAddr,
@@ -123,14 +124,19 @@ func (d *DoQClient) Resolve(q *dns.Msg, ci ClientInfo) (*dns.Msg, error) {
 	defer func() { q.Id = id }()
 
 	// Encode the query
-	b, err := q.Pack()
+	p, err := q.Pack()
 	if err != nil {
 		d.metrics.err.Add("pack", 1)
 		return nil, err
 	}
 
-	// Get a new stream in the session
-	stream, err := d.session.getStream()
+	// Add a length prefix
+	b := make([]byte, 2+len(p))
+	binary.BigEndian.PutUint16(b, uint16(len(p)))
+	copy(b[2:], p)
+
+	// Get a new stream in the connection
+	stream, err := d.connection.getStream()
 	if err != nil {
 		d.metrics.err.Add("getstream", 1)
 		return nil, err
@@ -138,7 +144,7 @@ func (d *DoQClient) Resolve(q *dns.Msg, ci ClientInfo) (*dns.Msg, error) {
 
 	// Write the query into the stream and close is. Only one stream per query/response
 	_ = stream.SetWriteDeadline(time.Now().Add(time.Second))
-	if _, err = stream.Write(b); err != nil {
+	if _, err = stream.Write(p); err != nil {
 		d.metrics.err.Add("write", 1)
 		return nil, err
 	}
@@ -147,10 +153,18 @@ func (d *DoQClient) Resolve(q *dns.Msg, ci ClientInfo) (*dns.Msg, error) {
 		return nil, err
 	}
 
-	// Read the response
 	_ = stream.SetReadDeadline(time.Now().Add(time.Second))
-	b, err = ioutil.ReadAll(stream)
-	if err != nil {
+
+	// DoQ requires a length prefix, like TCP
+	var length uint16
+	if err := binary.Read(stream, binary.BigEndian, &length); err != nil {
+		d.metrics.err.Add("read", 1)
+		return nil, err
+	}
+
+	// Read the response
+	b = make([]byte, length)
+	if _, err = io.ReadFull(stream, b); err != nil {
 		d.metrics.err.Add("read", 1)
 		return nil, err
 	}
@@ -180,7 +194,7 @@ func (d *DoQClient) String() string {
 	return d.id
 }
 
-type doqSession struct {
+type doqConnection struct {
 	hostname  string
 	endpoint  string
 	lAddr     net.IP
@@ -189,35 +203,35 @@ type doqSession struct {
 	log       *logrus.Entry
 	pool      *udpConnPool
 
-	session quic.Session
+	connection quic.Connection
 
 	mu sync.Mutex
 }
 
-func (s *doqSession) getStream() (quic.Stream, error) {
+func (s *doqConnection) getStream() (quic.Stream, error) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	// If we don't have a session yet, make one
-	if s.session == nil {
+	// If we don't have a connection yet, make one
+	if s.connection == nil {
 		var err error
-		s.session, err = quicDial(s.hostname, s.endpoint, s.lAddr, s.tlsConfig, s.config, s.pool)
+		s.connection, err = quicDial(s.hostname, s.endpoint, s.lAddr, s.tlsConfig, s.config, s.pool)
 		if err != nil {
-			s.log.WithError(err).Error("failed to open session")
+			s.log.WithError(err).Error("failed to open connection")
 			return nil, err
 		}
 	}
 
-	stream, err := s.session.OpenStream()
+	stream, err := s.connection.OpenStream()
 	if netErr, ok := err.(net.Error); ok && (netErr.Timeout() || netErr.Temporary()) {
-		// Try to open a new session
-		_ = s.session.CloseWithError(DOQNoError, "")
-		s.session, err = quicDial(s.hostname, s.endpoint, s.lAddr, s.tlsConfig, s.config, s.pool)
+		// Try to open a new connection
+		_ = s.connection.CloseWithError(DOQNoError, "")
+		s.connection, err = quicDial(s.hostname, s.endpoint, s.lAddr, s.tlsConfig, s.config, s.pool)
 		if err != nil {
-			s.log.WithError(err).Error("failed to open session")
+			s.log.WithError(err).Error("failed to open connection")
 			return nil, err
 		}
-		stream, err = s.session.OpenStream()
+		stream, err = s.connection.OpenStream()
 		if err != nil {
 			s.log.WithError(err).Error("failed to open stream")
 			return nil, err
