@@ -30,13 +30,13 @@ func (m MultiDB) Reload() (BlocklistDB, error) {
 	return NewMultiDB(newDBs...)
 }
 
-func (m MultiDB) Match(q dns.Question) (net.IP, string, string, bool) {
+func (m MultiDB) Match(q dns.Question) (net.IP, string, *BlocklistMatch, bool) {
 	for _, db := range m.dbs {
-		if ip, name, rule, ok := db.Match(q); ok {
-			return ip, name, rule, ok
+		if ip, name, match, ok := db.Match(q); ok {
+			return ip, name, match, ok
 		}
 	}
-	return nil, "", "", false
+	return nil, "", nil, false
 }
 
 func (m MultiDB) String() string {
