@@ -14,7 +14,15 @@ type BlocklistDB interface {
 
 	// Returns true if the question matches a rule. If the IP is not nil,
 	// respond with the given IP. NXDOMAIN otherwise.
-	Match(q dns.Question) (net.IP, string, string, bool)
+	Match(q dns.Question) (net.IP, string, *BlocklistMatch, bool)
 
 	fmt.Stringer
+}
+
+// BlocklistMatch is returned by blocklists when a match is found. It contains
+// information about what rule matched, what list it was from etc. Used mostly
+// for logging.
+type BlocklistMatch struct {
+	List string // Identifier or name of the blocklist
+	Rule string // Identifier for the rule that matched
 }
