@@ -13,7 +13,7 @@ func TestCidrDB(t *testing.T) {
 		"1.2.0.0/16",
 		"2a03:2880:f101:83::0/64",
 	})
-	db, err := NewCidrDB(loader)
+	db, err := NewCidrDB("testlist", loader)
 	require.NoError(t, err)
 
 	tests := []struct {
