@@ -14,9 +14,10 @@ import (
 	"sync"
 	"time"
 
+	"github.com/quic-go/quic-go"
+	"github.com/quic-go/quic-go/http3"
+
 	"github.com/jtacoma/uritemplates"
-	"github.com/lucas-clemente/quic-go"
-	"github.com/lucas-clemente/quic-go/http3"
 	"github.com/miekg/dns"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/net/http2"
@@ -253,11 +254,11 @@ func dohQuicTransport(endpoint string, opt DoHClientOptions) (http.RoundTripper,
 
 	// When using a custom dialer, we have to track/close connections ourselves
 	pool := new(udpConnPool)
-	dialer := func(network, addr string, tlsConfig *tls.Config, config *quic.Config) (quic.EarlySession, error) {
+	dialer := func(ctx context.Context, addr string, tlsConfig *tls.Config, config *quic.Config) (quic.EarlyConnection, error) {
 		return quicDial(u.Hostname(), addr, lAddr, tlsConfig, config, pool)
 	}
 	if opt.BootstrapAddr != "" {
-		dialer = func(network, addr string, tlsConfig *tls.Config, config *quic.Config) (quic.EarlySession, error) {
+		dialer = func(ctx context.Context, addr string, tlsConfig *tls.Config, config *quic.Config) (quic.EarlyConnection, error) {
 			_, port, err := net.SplitHostPort(addr)
 			if err != nil {
 				return nil, err
@@ -277,7 +278,7 @@ func dohQuicTransport(endpoint string, opt DoHClientOptions) (http.RoundTripper,
 	return &http3ReliableRoundTripper{tr, pool}, nil
 }
 
-func quicDial(hostname, rAddr string, lAddr net.IP, tlsConfig *tls.Config, config *quic.Config, pool *udpConnPool) (quic.EarlySession, error) {
+func quicDial(hostname, rAddr string, lAddr net.IP, tlsConfig *tls.Config, config *quic.Config, pool *udpConnPool) (quic.EarlyConnection, error) {
 	udpAddr, err := net.ResolveUDPAddr("udp", rAddr)
 	if err != nil {
 		return nil, err
@@ -290,7 +291,7 @@ func quicDial(hostname, rAddr string, lAddr net.IP, tlsConfig *tls.Config, confi
 	return quic.DialEarly(udpConn, udpAddr, hostname, tlsConfig, config)
 }
 
-// Wrapper for http3.RoundTripper due to https://github.com/lucas-clemente/quic-go/issues/765
+// Wrapper for http3.RoundTripper due to https://github.com/quic-go/quic-go/issues/765
 // This wrapper will transparently re-open expired connections. Should be removed once the issue
 // has been fixed upstream.
 type http3ReliableRoundTripper struct {
