@@ -5,6 +5,7 @@ import (
 	"net"
 
 	"github.com/miekg/dns"
+	"github.com/sirupsen/logrus"
 )
 
 // ECSModifier manipulates EDNS0 Client Subnet in queries.
@@ -17,7 +18,7 @@ type ECSModifier struct {
 var _ Resolver = &ECSModifier{}
 
 // ECSModifierFunc takes a DNS query and modifies its EDN0 Client Subdomain record
-type ECSModifierFunc func(q *dns.Msg, ci ClientInfo)
+type ECSModifierFunc func(id string, q *dns.Msg, ci ClientInfo)
 
 // NewECSModifier initializes an ECS modifier.
 func NewECSModifier(id string, resolver Resolver, f ECSModifierFunc) (*ECSModifier, error) {
@@ -33,7 +34,7 @@ func (r *ECSModifier) Resolve(q *dns.Msg, ci ClientInfo) (*dns.Msg, error) {
 
 	// Modify the query
 	if r.modifier != nil {
-		r.modifier(q, ci)
+		r.modifier(r.id, q, ci)
 	}
 
 	// Pass it on upstream
@@ -44,27 +45,35 @@ func (r *ECSModifier) String() string {
 	return r.id
 }
 
-func ECSModifierDelete(q *dns.Msg, ci ClientInfo) {
+func ECSModifierDelete(id string, q *dns.Msg, ci ClientInfo) {
 	edns0 := q.IsEdns0()
 	if edns0 == nil {
 		return
 	}
 	// Filter out any ECS options
+	var hasECS bool
 	newOpt := make([]dns.EDNS0, 0, len(edns0.Option))
 	for _, opt := range edns0.Option {
 		if _, ok := opt.(*dns.EDNS0_SUBNET); ok {
+			hasECS = true
 			continue
 		}
 		newOpt = append(newOpt, opt)
 	}
 	edns0.Option = newOpt
+
+	// Only log if id is set to a non-empty string. Avoids double-logging
+	// if called by other modifiers
+	if hasECS && id != "" {
+		logger(id, q, ci).Debug("removing ecs option")
+	}
 }
 
 func ECSModifierAdd(addr net.IP, prefix4, prefix6 uint8) ECSModifierFunc {
 
-	return func(q *dns.Msg, ci ClientInfo) {
+	return func(id string, q *dns.Msg, ci ClientInfo) {
 		// Drop any existing ECS options
-		ECSModifierDelete(q, ci)
+		ECSModifierDelete("", q, ci)
 
 		// If no address is configured, use that of the client
 		sourceIP := addr
@@ -102,17 +111,27 @@ func ECSModifierAdd(addr net.IP, prefix4, prefix6 uint8) ECSModifierFunc {
 		ecs.SourceScope = 0
 		ecs.Address = sourceIP
 		edns0.Option = append(edns0.Option, ecs)
+
+		logger(id, q, ci).WithFields(logrus.Fields{
+			"ecs":  sourceIP,
+			"mask": mask,
+		}).Debug("adding ecs option")
 	}
 }
 
 func ECSModifierPrivacy(prefix4, prefix6 uint8) ECSModifierFunc {
-	return func(q *dns.Msg, ci ClientInfo) {
+	return func(id string, q *dns.Msg, ci ClientInfo) {
 		edns0 := q.IsEdns0()
 		if edns0 == nil {
 			return
 		}
 
 		// Find the ECS option
+		var (
+			hasECS     bool
+			beforeAddr net.IP
+			afterAddr  net.IP
+		)
 		for _, opt := range edns0.Option {
 			ecs, ok := opt.(*dns.EDNS0_SUBNET)
 			if !ok {
@@ -120,14 +139,26 @@ func ECSModifierPrivacy(prefix4, prefix6 uint8) ECSModifierFunc {
 			}
 			switch ecs.Family {
 			case 1: // ip4
-				addr := ecs.Address.To4()
-				ecs.Address = addr.Mask(net.CIDRMask(int(prefix4), 32))
+				beforeAddr = ecs.Address.To4()
+				afterAddr = beforeAddr.Mask(net.CIDRMask(int(prefix4), 32))
+				ecs.Address = afterAddr
 				ecs.SourceNetmask = prefix4
 			case 2: // ip6
-				addr := ecs.Address
-				ecs.Address = addr.Mask(net.CIDRMask(int(prefix6), 128))
+				beforeAddr = ecs.Address
+				afterAddr = beforeAddr.Mask(net.CIDRMask(int(prefix6), 128))
+				ecs.Address = afterAddr
 				ecs.SourceNetmask = prefix6
 			}
+			hasECS = true
+		}
+
+		if hasECS {
+			logger(id, q, ci).WithFields(logrus.Fields{
+				"ip4prefix":   prefix4,
+				"ip6prefix":   prefix6,
+				"before-addr": beforeAddr,
+				"after-addr":  afterAddr,
+			}).Debug("modifying ecs privacy")
 		}
 	}
 }
