@@ -13,19 +13,22 @@ import (
 )
 
 type route struct {
-	types    []uint16
-	class    uint16
-	name     *regexp.Regexp
-	source   *net.IPNet
-	weekdays []time.Weekday
-	before   *TimeOfDay
-	after    *TimeOfDay
-	inverted bool // invert the matching behavior
-	resolver Resolver
+	types         []uint16
+	class         uint16
+	name          *regexp.Regexp
+	source        *net.IPNet
+	weekdays      []time.Weekday
+	before        *TimeOfDay
+	after         *TimeOfDay
+	inverted      bool // invert the matching behavior
+	dohPath       *regexp.Regexp
+	resolver      Resolver
+	listenerID    *regexp.Regexp
+	tlsServerName *regexp.Regexp
 }
 
 // NewRoute initializes a route from string parameters.
-func NewRoute(name, class string, types, weekdays []string, before, after, source string, resolver Resolver) (*route, error) {
+func NewRoute(name, class string, types, weekdays []string, before, after, source, dohPath, listenerID, tlsServerName string, resolver Resolver) (*route, error) {
 	if resolver == nil {
 		return nil, errors.New("no resolver defined for route")
 	}
@@ -53,6 +56,18 @@ func NewRoute(name, class string, types, weekdays []string, before, after, sourc
 	if err != nil {
 		return nil, err
 	}
+	dohRe, err := regexp.Compile(dohPath)
+	if err != nil {
+		return nil, err
+	}
+	listenerRe, err := regexp.Compile(listenerID)
+	if err != nil {
+		return nil, err
+	}
+	tlsRe, err := regexp.Compile(tlsServerName)
+	if err != nil {
+		return nil, err
+	}
 	var sNet *net.IPNet
 	if source != "" {
 		_, sNet, err = net.ParseCIDR(source)
@@ -61,14 +76,17 @@ func NewRoute(name, class string, types, weekdays []string, before, after, sourc
 		}
 	}
 	return &route{
-		types:    t,
-		class:    c,
-		name:     re,
-		weekdays: w,
-		before:   b,
-		after:    a,
-		source:   sNet,
-		resolver: resolver,
+		types:         t,
+		class:         c,
+		name:          re,
+		weekdays:      w,
+		before:        b,
+		after:         a,
+		source:        sNet,
+		dohPath:       dohRe,
+		listenerID:    listenerRe,
+		tlsServerName: tlsRe,
+		resolver:      resolver,
 	}, nil
 }
 
@@ -84,6 +102,15 @@ func (r *route) match(q *dns.Msg, ci ClientInfo) bool {
 		return r.inverted
 	}
 	if r.source != nil && !r.source.Contains(ci.SourceIP) {
+		return r.inverted
+	}
+	if !r.dohPath.MatchString(ci.DoHPath) {
+		return r.inverted
+	}
+	if !r.listenerID.MatchString(ci.Listener) {
+		return r.inverted
+	}
+	if !r.tlsServerName.MatchString(ci.TLSServerName) {
 		return r.inverted
 	}
 	if len(r.weekdays) > 0 || r.before != nil || r.after != nil {
@@ -119,9 +146,48 @@ func (r *route) Invert(value bool) {
 
 func (r *route) String() string {
 	if r.isDefault() {
-		return fmt.Sprintf("default->%s", r.resolver)
+		return "(default)"
 	}
-	return fmt.Sprintf("%s->%s", r.name, r.resolver)
+	var fragments []string
+	if len(r.types) > 0 {
+		var types []string
+		for _, t := range r.types {
+			types = append(types, dns.TypeToString[t])
+		}
+		fragments = append(fragments, fmt.Sprintf("types=%v", types))
+	}
+	if r.name.String() != "" {
+		fragments = append(fragments, "name="+r.name.String())
+	}
+	if r.class != 0 {
+		s, _ := classToString(r.class)
+		fragments = append(fragments, "class="+s)
+	}
+	if r.source != nil {
+		fragments = append(fragments, "source="+r.source.String())
+	}
+	if r.dohPath.String() != "" {
+		fragments = append(fragments, "doh-path="+r.dohPath.String())
+	}
+	if r.listenerID.String() != "" {
+		fragments = append(fragments, "listener="+r.listenerID.String())
+	}
+	if r.tlsServerName.String() != "" {
+		fragments = append(fragments, "servername="+r.tlsServerName.String())
+	}
+	if len(r.weekdays) > 0 {
+		fragments = append(fragments, fmt.Sprintf("weekdays=%v", r.weekdays))
+	}
+	if r.after != nil {
+		fragments = append(fragments, "after="+r.after.String())
+	}
+	if r.before != nil {
+		fragments = append(fragments, "before="+r.before.String())
+	}
+	if r.inverted {
+		fragments = append(fragments, "invert=true")
+	}
+	return "(" + strings.Join(fragments, ",") + ")"
 }
 
 func (r *route) isDefault() bool {
@@ -159,7 +225,7 @@ loop:
 	return types, nil
 }
 
-// Convert a DNS class string into is numerical form, for example "INET" -> 1.
+// Convert a DNS class string into its numerical form, for example "INET" -> 1.
 func stringToClass(s string) (uint16, error) {
 	switch strings.ToUpper(s) {
 	case "":
@@ -176,6 +242,24 @@ func stringToClass(s string) (uint16, error) {
 		return 255, nil
 	default:
 		return 0, fmt.Errorf("unknown class '%s'", s)
+	}
+}
+
+// Convert a DNS class identifier into its string from, for example 1 -> "INET".
+func classToString(class uint16) (string, error) {
+	switch class {
+	case 1:
+		return "IN", nil
+	case 3:
+		return "CH", nil
+	case 4:
+		return "HS", nil
+	case 254:
+		return "NONE", nil
+	case 255:
+		return "ANY", nil
+	default:
+		return "", fmt.Errorf("unknown class identifier %d", class)
 	}
 }
 
@@ -238,4 +322,8 @@ func (t *TimeOfDay) isBefore(hour, minute int) bool {
 
 func (t *TimeOfDay) isAfter(hour, minute int) bool {
 	return t.hour > hour || (t.hour == hour && t.minute > minute)
+}
+
+func (t *TimeOfDay) String() string {
+	return fmt.Sprintf("%2d:%2d", t.hour, t.minute)
 }
