@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/miekg/dns"
+	"github.com/sirupsen/logrus"
 )
 
 // Router for DNS requests based on query type and/or name. Implements the Resolver interface.
@@ -56,7 +57,10 @@ func (r *Router) Resolve(q *dns.Msg, ci ClientInfo) (*dns.Msg, error) {
 		if !route.match(q, ci) {
 			continue
 		}
-		log.WithField("resolver", route.resolver.String()).Debug("routing query to resolver")
+		log.WithFields(logrus.Fields{
+			"route":    route.String(),
+			"resolver": route.resolver.String()},
+		).Debug("routing query to resolver")
 		r.metrics.route.Add(route.resolver.String(), 1)
 		a, err := route.resolver.Resolve(q, ci)
 		if err != nil {
