@@ -1,8 +1,8 @@
 package rdns
 
 var (
-	BuildVersion string = "v0.1.0"
-	BuildTime    string = ""
-	BuildNumber  string = "001"
+	BuildVersion string = "v0.1.21"
+	BuildTime    string = "Fri Feb  3 00:24:18 UTC 2023"
+	BuildNumber     string = "4"
 )
 
